pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gitgpt"
version = "0.1.0"
authors = [
  { name="Your Name", email="your@email.com" },
]
# <-- YOU MUST HAVE A NEWLINE HERE
description = "A natural language to git command translator."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
# <-- YOU MUST HAVE A NEWLINE HERE
dependencies = [
    "google-generativeai",
    "python-dotenv",
    "questionary",
    "rich",
]

[project.urls]
"Homepage" = "https://github.com/YOUR_USERNAME/gitgpt-cli" # Optional

[project.scripts]
gitgpt = "gitgpt.main:run_cli"